requirements.txt
# Core dependencies for Lambda-compatible version
boto3==1.38.23
pinecone==7.0.1
pydantic==2.11.5
python-dotenv==1.1.0

# Lightweight web scraping (Lambda-friendly)
requests==2.32.3
beautifulsoup4==4.13.4
lxml==5.4.0
certifi>=2023.0.0

# Optional: Full crawl4ai (for local development)
# crawl4ai==0.6.3

# Optional: Neo4j (for local graph storage)
# neo4j==5.28.1

# Web UI
flask==3.0.0
flask-login==0.6.3
email-validator==2.2.0

# Optional: Streamlit (for local UI)
# streamlit==1.45.1